requirements.txt
ollama
pyttsx3
SpeechRecognition
Pillow
pyaudio
gtts